data and data cleaning for 401(k) example
There are 12 files corresponding to trim=0:1 and quintile=0:5

trim=0 means all observations are kept. trim=1 means that untreated observations with extreme propensities (outside the range of treated propensities) are dropped as in https://arxiv.org/pdf/1802.08667.pdf. This amounts to dropping 67 observations

quintile=0 means the entire population. quintile=1 means the subpopulation whose income is in the first quintile (and so on). These localized / subgroup ATEs are also reported in https://arxiv.org/pdf/1802.08667.pdf